+++ lib/lastfm_archive/archive/transformers/file_archive_transformer.ex
@@ -6,19 +6,21 @@ defmodule LastfmArchive.Archive.Transformers.FileArchiveTransformer do
   @behaviour LastfmArchive.Behaviour.Transformer
 
   alias Explorer.DataFrame
+  alias LastfmArchive.Archive.DerivedArchive
   alias LastfmArchive.Archive.Metadata
+  alias LastfmArchive.Archive.Transformers.FileArchiveTransformerSettings
   alias LastfmArchive.Behaviour.Archive
 
-  import LastfmArchive.Utils, only: [create_dir: 2, create_filepath: 2, month_range: 2, year_range: 1, write: 2]
+  use LastfmArchive.Behaviour.DataFrameIo, formats: FileArchiveTransformerSettings.available_formats()
+
+  import LastfmArchive.Utils, only: [create_dir: 2, create_filepath: 3, month_range: 2, year_range: 1, write: 2]
   require Logger
 
-  @data_frame_io Application.compile_env(:lastfm_archive, :data_frame_io, DataFrame)
   @file_io Application.compile_env(:lastfm_archive, :file_io, Elixir.File)
-  @formats [:tsv, :parquet]
 
   @impl true
-  def apply(%{creator: user} = metadata, [format: format] = opts) do
-    :ok = create_dir(user, format: format)
+  def apply(%{creator: user} = metadata, opts) do
+    :ok = create_dir(user, format: Keyword.fetch!(opts, :format))
     :ok = transform(metadata, year_range(metadata.temporal) |> Enum.to_list(), opts)
 
     {:ok, %{metadata | modified: DateTime.utc_now()}}
@@ -33,15 +35,26 @@ defmodule LastfmArchive.Archive.Transformers.FileArchiveTransformer do
 
   defp transform(%Metadata{creator: user} = metadata, [%Date{year: year} | _] = months, opts) do
     format = Keyword.get(opts, :format)
+    overwrite = Keyword.get(opts, :overwrite, false)
 
-    case create_filepath(user, "#{format}/#{year}.#{format}.gz") do
+    case create_filepath(user, format, "#{format}/#{year}.#{format}") do
       {:ok, filepath} ->
-        Logger.info("\nCreating #{format} file for #{year} scrobbles.")
-        :ok = create_dataframe(metadata, months) |> write_data_frame(filepath, format: format)
+        transform({metadata, year, months, filepath, format})
 
-      {:error, :file_exists} ->
-        Logger.info("\n#{format} file exists, skipping #{year} scrobbles.")
+      {:error, :file_exists, filepath} ->
+        maybe_skip_transform({metadata, year, months, filepath, format}, overwrite: overwrite)
     end
+  end
+
+  defp transform({metadata, year, months, filepath, format}) do
+    Logger.info("\nCreating #{format} file for #{year} scrobbles.")
+    :ok = create_dataframe(metadata, months) |> write_data_frame(filepath, format: format)
+  end
+
+  defp maybe_skip_transform(args, overwrite: true), do: transform(args)
+
+  defp maybe_skip_transform({_, year, _, _, format}, overwrite: false) do
+    Logger.info("\n#{format} file exists, skipping #{year} scrobbles.")
   end
 
   defp create_dataframe(metadata, months) do
@@ -52,18 +65,24 @@ defmodule LastfmArchive.Archive.Transformers.FileArchiveTransformer do
     |> DataFrame.concat_rows()
   end
 
-  defp write_data_frame(df, filepath, format: format) when format in @formats do
-    {format, opts} = if format == :tsv, do: {:csv, [delimiter: "\t"]}, else: {format, []}
-
-    write_fun = fn df ->
-      df
-      |> DataFrame.collect()
-      |> dump_data_frame({format, opts})
-      |> then(fn data -> @file_io.write(filepath, data, [:compressed]) end)
-    end
-
-    write(df, write_fun)
+  defp write_data_frame(df, filepath, format: format) do
+    write(df, write_fun(filepath, format))
   end
 
-  defp dump_data_frame(df, {format, opts}), do: apply(@data_frame_io, :"dump_#{format}!", [df, opts])
+  defp write_fun(filepath, format) when format == :csv do
+    fn df ->
+      df
+      |> DataFrame.collect()
+      |> dump_data_frame(format, DerivedArchive.write_opts(format))
+      |> then(fn data -> @file_io.write(filepath, data, [:compressed]) end)
+    end
+  end
+
+  defp write_fun(filepath, format) do
+    fn df ->
+      df
+      |> DataFrame.collect()
+      |> to_data_frame(filepath, format, DerivedArchive.write_opts(format))
+    end
+  end
 end